Add functionality for user to need an address to checkout.

+++ app/controllers/user/orders_controller.rb
@@ -10,18 +10,23 @@ class User::OrdersController < ApplicationController
   end
 
   def create
-    order = current_user.orders.new
-    order.save
-      cart.items.each do |item|
-        order.order_items.create({
-          item: item,
-          quantity: cart.count_of(item.id),
-          price: item.price
-          })
-      end
-    session.delete(:cart)
-    flash[:notice] = "Order created successfully!"
-    redirect_to '/profile/orders'
+    if current_user.has_no_addresses?
+      flash[:error] = 'You must add an address to checkout!'
+      redirect_to '/profile/addresses/new'
+    else
+      order = current_user.orders.new
+      order.save
+        cart.items.each do |item|
+          order.order_items.create({
+            item: item,
+            quantity: cart.count_of(item.id),
+            price: item.price
+            })
+        end
+      session.delete(:cart)
+      flash[:notice] = "Order created successfully!"
+      redirect_to '/profile/orders'
+    end
   end
 
   def cancel
